feat(epg): rework epg to work offline (#391)

+++ shared/src/commonMain/kotlin/fr/outadoc/justchatting/feature/chat/data/pubsub/plugin/viewercount/PubSubViewerCountMessage.kt
@@ -10,6 +10,6 @@ internal sealed class PubSubViewerCountMessage {
     @SerialName("viewcount")
     data class ViewCount(
         @SerialName("viewers")
-        val viewers: Int,
+        val viewers: Long,
     ) : PubSubViewerCountMessage()
 }
